==> r_scripts/r_scripts_treated_worms/1_package_info.Rmd <==
---
title: "1_package_info"
author: "Hailey Sounart"
date: "10/20/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(include = TRUE, echo = TRUE, message = TRUE, warning =FALSE)

```

## Load all packages used in analysis

```{r load packages}

  library(STutility)
  library(magrittr)
  library(ggplot2)
  library(magick)
  library(grid)
  library(openxlsx)
  library(ggpubr)

Sys.time()
```


## Get session information

```{r, include = TRUE, echo = TRUE, message = TRUE}

sessioninfo::session_info()

Sys.time()
```
Output is the versions of all package dependencies for running the R analysis scripts 2_add_metadata_treated_worms.Rmd, 3_quality_check_filtering_treated_worms.Rmd, 4_normalization_treated_worms.Rmd, 6_look_at_treated_analysis.Rmd, and 7_genes_on_tissue_visualization_treated_worms.Rmd.
