==> r_scripts/r_scripts_treated_worms/6_look_at_treated_analysis.Rmd <==
---
title: "6_look_at_treated_analysis"
author: "Hailey Sounart"
date: "3/7/2023"
output: html_document
---

````{r, include = TRUE, echo = TRUE, message = TRUE}

start_time <- Sys.time()
start_time
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(include = TRUE, echo = TRUE, message = FALSE, warning =FALSE)
```

```{r load packages}
  library(STutility)
  library(ggplot2)
  library(openxlsx)
```

```{r set working directory}

setwd("<path to working directory>")

``` 

# Read in Fold Changes

``` {r read in fold changes}

genes_of_interest_fc <- readRDS("genes_of_interest_fc.rds")
genes_of_interest_fc 

genes_of_interest_wol_fc <- readRDS("genes_of_interest_wol_fc.rds")
genes_of_interest_wol_fc 

```

# Make into Supp Table format

Rename rows/genes
``` {r rename genes}

genes_of_interest_fc_renamed <- genes_of_interest_fc
genes_of_interest_fc_renamed_genes <- gsub("bmalayi-", "", rownames(genes_of_interest_fc_renamed))
rownames(genes_of_interest_fc_renamed) <- genes_of_interest_fc_renamed_genes
genes_of_interest_fc_renamed

genes_of_interest_wol_fc_renamed <- genes_of_interest_wol_fc
genes_of_interest_wol_fc_renamed_genes <- gsub("bmalayi-", "", rownames(genes_of_interest_wol_fc_renamed))
rownames(genes_of_interest_wol_fc_renamed) <- genes_of_interest_wol_fc_renamed_genes
genes_of_interest_wol_fc_renamed


```

Read in annotations
```{r add annotations}

#read in the bmalayi annotation file as a table
all_bmalayi_gene_annotations <- read.table("brugia_malayi.gene_biotype_product.map.tsv", sep="\t", header=T, stringsAsFactors = F)
all_bmalayi_gene_annotations

```

Add annotations into table
```{r add annotations 2}

#all_bmalayi_gene_annotations$gene_id
all_bmalayi_gene_annotations_table <- all_bmalayi_gene_annotations
rownames(all_bmalayi_gene_annotations_table) <- all_bmalayi_gene_annotations$gene_id
all_bmalayi_gene_annotations_table

all_bmalayi_gene_annotations_table <- all_bmalayi_gene_annotations_table[ , -which(names(all_bmalayi_gene_annotations_table) %in% "gene_id")]
all_bmalayi_gene_annotations_table

```

Merge annotations into gene table
```{r add annotations 3}

genes_of_interest_fc_renamed_annot <- transform(merge(genes_of_interest_fc_renamed, all_bmalayi_gene_annotations_table,by="row.names",all.x=TRUE), row.names=Row.names, Row.names=NULL) #,all.y=TRUE
genes_of_interest_fc_renamed_annot

genes_of_interest_wol_fc_renamed_annot <- transform(merge(genes_of_interest_wol_fc_renamed, all_bmalayi_gene_annotations_table,by="row.names",all.x=TRUE), row.names=Row.names, Row.names=NULL) #,all.y=TRUE
genes_of_interest_wol_fc_renamed_annot

``` 

## Save the gene lists
```{r save colocal genes}

# Supplementary Data 9

saveRDS(genes_of_interest_fc_renamed_annot, file = "Supplementary_Data_9A.rds")
write.xlsx(genes_of_interest_fc_renamed_annot, file = "Supplementary_Data_9A.xlsx", col.names=TRUE, row.names=TRUE)

saveRDS(genes_of_interest_wol_fc_renamed_annot, file = "Supplementary_Data_9B.rds")
write.xlsx(genes_of_interest_wol_fc_renamed_annot, file = "Supplementary_Data_9B.xlsx", col.names=TRUE, row.names=TRUE)

``` 

## Plot - Figure 6E

### Read in Normalized object

``` {r read norm object}

bmalay.treated_norm_obj <- readRDS("bmalay.treated.normalized.obj.rds")
bmalay.treated_norm_obj 

```

### Add wolbachia metadata to brugia object

```{r add wol}

bmalay.treated_norm_obj@meta.data$wolbachia <- ifelse((bmalay.treated_norm_obj@meta.data$counts.wBm >= 1), 1, 0)
bmalay.treated_norm_obj@meta.data

```

```{r text genes}

wol_treated_text_genes <- c("bmalayi-WBGene00231090", "bmalayi-WBGene00234242", "bmalayi-WBGene00230716", "bmalayi-WBGene00233153", "bmalayi-WBGene00227756","bmalayi-WBGene00234295", "bmalayi-WBGene00225880", "bmalayi-WBGene00233912", "bmalayi-WBGene00221322", "bmalayi-WBGene00226491","bmalayi-WBGene00220387", "bmalayi-WBGene00224725", "bmalayi-WBGene00222232", "bmalayi-WBGene00224389")
wol_treated_text_genes

```

### Dotplot
```{r dotplot}

merge_wol_dotplot <- DotPlot(bmalay.treated_norm_obj, assay = "SCT", features = wol_treated_text_genes, cols ="RdBu",  group.by = "wolbachia", split.by = "condition", dot.scale = 10) + RotatedAxis() + scale_x_discrete(labels = c("bmalayi-WBGene00231090" = "Bma-cul-2", "bmalayi-WBGene00234242" = "Bm13981","bmalayi-WBGene00230716" = "Bm10455", "bmalayi-WBGene00233153" = "Bm12892", "bmalayi-WBGene00227756" = "Bma-try-1", "bmalayi-WBGene00234295" = "Bm14034", "bmalayi-WBGene00225880" = "Bma-cyb-3", "bmalayi-WBGene00233912" = "Bma-ubq-2", "bmalayi-WBGene00221322" = "Bm1061", "bmalayi-WBGene00226491" = "Bma-lap-2", "bmalayi-WBGene00220387" = "Bm126", "bmalayi-WBGene00224725" = "Bma-glb-19", "bmalayi-WBGene00222232" = "Bma-snx-1","bmalayi-WBGene00224389" = "Bma-ned-8")) + scale_y_discrete(labels = c("1_doxycycline" = "W+ T", "1_control" = "W+ C", "0_doxycycline" = "W- T", "0_control" = "W- C")) + xlab("Genes") + ylab("Condition")
merge_wol_dotplot

ggsave(file = paste0("Figure_6E.pdf"), plot = merge_wol_dotplot, width = 10, height =5, dpi = 300)

```

```{r, include = TRUE, echo = TRUE, message = TRUE}

end_time <- Sys.time()
end_time
```

```{r, include = TRUE, echo = TRUE, message = TRUE}

elapsed_time <- round((end_time-start_time), 3)
elapsed_time
```

Outputs are excel files containing the genes of interest and their fold change values, presented in Supplementary Data 9. The output dotplot is presented in Figure 6E.
